Change hxtool links to point to my hxtool fork
At this time, the dev branch of the official hxtool repository contains about a year's worth of unreleased improvements. It holds 66 commits in total, 26 of which are ahead of master. I tried to raise the matter of merging dev into master in the discussion of the auto-detect PR, but the hxtool maintainer didn't really respond to that. There's no telling when those improvements will be merged and officially released as part of hxtool; it could take months, for all I know.

The fork contains additional improvements such as the poke and nav commands, for which the situation is similar.

The GitHub UI makes it easy to reach a fork's upstream repository. The fork's readme has prominent links back to the upstream, too.

Therefore, it seems beneficial for users to point them at the fork instead.